Initial work on logic flow overhaul
Basically, there are two ways to run the logic flow in a game engine
like this, and I think I picked the wrong one, wanting to promote
developer control. Either implement a `frame_sync` function and give the
user complete logical control over the engine, or expect the user to
expose `init` and `update` functions that manipulate a global state,
which the engine hooks into a calls whenever it wants. This engine began
with the former, but is transitioning to the latter. Both approaches
have upsides and downsides.